Address Arch discussion feedback
